+++ src/rpizero2/src/multicoremanager.h
@@ -22,9 +22,6 @@
 #include <circle/types.h>
 #include <circle/usb/usbgamepad.h>
 #include <circle/usb/usbhcidevice.h>
-#include <vc4/sound/vchiqsoundbasedevice.h>
-#include <vc4/sound/vchiqsounddevice.h>
-#include <vc4/vchiq/vchiqdevice.h>
 
 enum class CoreStatus {
     Init,